Changed a lot of stuff
Disconnected visuals and the logic by creating a new maths class.

Separated functions dictionary to a separate class and combined it with the function class.

Converted coefficient and exponents data types from an integer to a float.

Rather than throw an error when an input isn't assigned but should be, just make the input x instead.

Created functions to get term variables from another class.

Created a function to take in an equation, and convert it into computer maths notation.